+++ Sources/KeyboardKit/_Deprecated/KeyboardInputTextComponent+Deprecated.swift
@@ -1,0 +1,6 @@
+#if os(iOS) || os(tvOS)
+import Foundation
+
+@available(*, deprecated, renamed: "KeyboardInputComponent")
+public typealias KeyboardInputTextComponent = KeyboardInputComponent
+#endif